test: colocate real-only Redis suites with classic tests via realRedisTest
