+++ tests/invalidation/RedisInvalidationBus.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest'
 import { RedisInvalidationBus } from '../../src/invalidation/RedisInvalidationBus'
-import { createTestRedis } from '../helpers/test-redis'
+import { createTestRedis, realRedisTest } from '../helpers/test-redis'
 
 describe('RedisInvalidationBus', () => {
   afterEach(() => {
@@ -424,5 +424,38 @@ describe('RedisInvalidationBus', () => {
     await unsubscribe()
     publisher.disconnect()
     subscriber.disconnect()
+  })
+
+  realRedisTest.it('propagates messages across independent bus instances', async () => {
+    const publisher = createTestRedis()
+    const subscriber = createTestRedis()
+    const channel = 'bus:test'
+
+    const busA = new RedisInvalidationBus({ publisher, subscriber: publisher.duplicate(), channel })
+    const busB = new RedisInvalidationBus({
+      publisher: subscriber,
+      subscriber: subscriber.duplicate(),
+      channel
+    })
+
+    const receivedByA: unknown[] = []
+    const receivedByB: unknown[] = []
+
+    const unsubA = await busA.subscribe(async (msg) => {
+      receivedByA.push(msg)
+    })
+    const unsubB = await busB.subscribe(async (msg) => {
+      receivedByB.push(msg)
+    })
+
+    await busB.publish({ scope: 'keys', sourceId: 'b', keys: ['k1', 'k2'], operation: 'invalidate' })
+
+    await vi.waitFor(() => {
+      expect(receivedByA).toHaveLength(1)
+      expect(receivedByB).toHaveLength(1)
+    })
+
+    await unsubA()
+    await unsubB()
   })
 })
