+++ vitest.config.ts
@@ -3,9 +3,7 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     setupFiles: ['./tests/setup.ts'],
-    // tests/integration/redis/* need a live Redis server (globalSetup in
-    // vitest.integration.config.ts), so they only run via that config.
-    exclude: ['**/node_modules/**', '**/dist/**', '**/docs-web/**', 'tests/integration/redis/**'],
+    exclude: ['**/node_modules/**', '**/dist/**', '**/docs-web/**'],
     coverage: {
       provider: 'v8',
       reporter: ['text', 'lcov', 'html'],
